tenstorrent: bh_arc: init ethernet with priority 15
Initialize ethernet with priorith 15. Note, this preserves the relative
init order since InitMrisc() is run with priority 14.
